Rename blockchain_storage class to Blockchain
Recursively replaced all instances of the string "blockchain_storage"
in src/ with the string "Blockchain".  This will result in some
functions having upper-case "Blockchain" in the function name, but that
can happily be changed if someone takes issue with it.

blockchain_storage.{h,cpp} and blockchain_storage_boost_serialization.h
are left unchanged.  blockchain_storage_boost_serialization.h is removed
as a dependency for Blockchain.{h,cpp} as that serialization method will
no longer be used.