Added tx_unlocked and get_output{,_global} functions to BlockchainDB
Rather than fetch a transaction to see if it's unlocked, added a
function to BlockchainDB for that, in case at a later time it is decided
to store unlock times outside of transactions for efficiency.

It is necessary to be able to get transaction outputs directly, and
currently those are indexed both by a global "all outputs that exist"
index and by a local "this is the offset for that output in the
transaction it was included in" index.