changed block adding and popping function calls
Blocks don't actually have the transactions inside them, but rather just
the transaction hashes, so add_block will now take a list of
transactions and pop_block will return a list of transactions.  Given
the other metadata being supplied to add_block, it might be more correct
to create a new struct "full_block" or similar that has said metadata
and the full transactions.  Revisit in the future.