Large commit -- redo block_to_main_chain with BlockchainDB
This commit is a rework of the handle_block_to_main_chain function to
bring it in line with BlockchainDB.  It also modifies various functions
that are called therein.

As this is an incredibly large commit, but all related things, putting
it in multiple commits seemed strange but fully documenting all changes
involved here in the commit message would be unwieldy.  The changes are
well-documented in the comments in the source, however!